Deduplicate platform test helper
